requirements.txt
cornice
requests
zope.sqlalchemy
